refactor and tidy Admiral agent lifecycle
