+++ commercial/app/jobs/AdmiralLifecycle.scala
@@ -1,11 +1,12 @@
-package commercial
+package jobs
 
+import agents.AdmiralAgent
 import app.LifecycleComponent
 import common.{JobScheduler, PekkoAsync}
 import play.api.inject.ApplicationLifecycle
 
-import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration._
+import scala.concurrent.{ExecutionContext, Future}
 
 class AdmiralLifecycle(
     appLifecycle: ApplicationLifecycle,
@@ -24,7 +25,7 @@ class AdmiralLifecycle(
   override def start(): Unit = {
     jobs.deschedule("AdmiralAgentRefreshJob")
 
-    jobs.scheduleEvery("AdmiralAgentRefreshJob", 10.seconds) {
+    jobs.scheduleEvery("AdmiralAgentRefreshJob", 2.hours) {
       admiralAgent.refresh()
     }
